requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.9.3            # via flask-migrate
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask, pip-tools
first==2.0.1              # via pip-tools
flake8==3.3.0
flask-migrate==2.0.4
flask-script==2.0.5       # via flask-migrate
flask-sqlalchemy==2.2
flask==0.12.2
gunicorn==19.7.1
idna==2.5                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
mako==1.0.7               # via alembic
markupsafe==1.0           # via jinja2, mako
mccabe==0.6.1             # via flake8
pip-tools==1.9.0
psycopg2==2.7.2
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
python-dateutil==2.6.1    # via alembic
python-editor==1.0.3      # via alembic
pyyaml==3.12
rauth==0.7.3
requests==2.18.1
six==1.10.0               # via pip-tools, python-dateutil
sqlalchemy==1.1.11        # via alembic, flask-sqlalchemy
urllib3==1.21.1           # via requests
werkzeug==0.12.2          # via flask
